Move file to common folder
